Quality: OOM retry, augmentation UI, SOAP FP32 moments, EMA load fix, defensive guards
- OOM retry in training loop: catch torch.cuda.OutOfMemoryError, clear VRAM,
  reset accumulation window, skip batch instead of crashing the session
- ScheduleFree eval/train wrapper for on-demand saves (stop/pause/manual)
- Color augmentation UI: brightness, contrast, saturation, hue, rotation
  fields in Dataset tab, wired to build_config/apply_config
- Loss function UI: MSE/Huber/Smooth L1 selector + Huber delta in Advanced tab
- Multires noise discount UI spinner in Noise Parameters section
- Conv Alpha UI spinner (was missing, only had hasattr fallback)
- SOAP optimizer: moments (exp_avg, exp_avg_sq) initialized in FP32 instead
  of param dtype to prevent BF16 precision loss in second moment accumulation
- SOAP _rotate: handle dtype mismatch between Q (BF16) and tensor (FP32)
- EMA load_state_dict: cast loaded shadow params to FP32 to maintain invariant
- Config fingerprint: skip keys absent from older checkpoints to avoid false
  SEVERE warnings on resume
- Path prefix check: use Path.is_relative_to() instead of string startswith
- Fixed DoRA/rsLoRA row overlap in LoRA grid layout
- 26 new unit tests covering collate, EMA precision, SOAP moments, rolling
  backup, config fingerprint, augmentation defaults, path checks, OOM reset
- Fixed pre-existing EMA test expectations for decay warmup formula

+++ dataset_sorter/ema.py
@@ -130,7 +130,7 @@ class EMAModel:
             )
         for i in range(min(len(saved), len(self.shadow_params))):
             if saved[i].shape == self.shadow_params[i].shape:
-                self.shadow_params[i] = saved[i].clone()
+                self.shadow_params[i] = saved[i].clone().float()
             else:
                 log.warning(
                     f"EMA param {i}: shape mismatch {saved[i].shape} vs "
